My project for Advanced Computer Graphics.
I was mainly the one in charge of rendering.
Use Ogre, CEGUI, Caleum, Hydrid.
Extracted models for World of Warcraft.

Just for fun. Maybe this project 's most difficult thing
is not about how to implemnt it , it's the different version of
Ogre, Caleum and Hydrid, how to get them work together that is
really time consuming.

Also thanks for Jia Song who wrote the calculation of satellites.